[Rename] share로 변경
